Constructing new classes/methods for DataMap
